Move event API views into their own file
