+++ hackathon_site/event/api_urls.py
@@ -1,9 +1,9 @@
 from django.urls import path
 
-from event import views
+from event import api_views
 
 app_name = "event"
 
 urlpatterns = [
-    path("users/user/", views.CurrentUserAPIView.as_view(), name="current-user"),
+    path("users/user/", api_views.CurrentUserAPIView.as_view(), name="current-user"),
 ]
