+++ hackathon_site/event/urls.py
@@ -1,5 +1,6 @@
 from django.contrib.auth import views as auth_views
 from django.urls import path
+from django.views.generic.base import TemplateView
 from event.views import IndexView
 
 app_name = "event"
@@ -11,4 +12,5 @@ urlpatterns = [
         auth_views.LoginView.as_view(template_name="event/login.html"),
         name="login",
     ),
+    path("dashboard/", TemplateView.as_view(template_name="event/dashboard_base.html"), name="dashboard")
 ]
